[Refactor] Onboarding View Process done
