+++ BE/Source/Onboarding/OnboardingLastView.swift
@@ -20,49 +20,46 @@ struct OnboardingLastView: View {
     """
     
     var body: some View {
-        ZStack {
-            Color.red.ignoresSafeArea()
-            VStack {
-                VStack(alignment: .leading, spacing: 8) {
-                    HStack {
-                        Text(text)
-                            .font(.title3)
-                            .fontWeight(.semibold)
-                            .foregroundColor(.white)
-                            .multilineTextAlignment(.leading)
+        VStack {
+            VStack(alignment: .leading, spacing: 8) {
+                HStack {
+                    Text(text)
+                        .font(.title3)
+                        .fontWeight(.semibold)
+                        .foregroundColor(.white)
+                        .multilineTextAlignment(.leading)
                         .lineSpacing(8)
-                        Spacer()
-                    }
                     Spacer()
-                    
-                }// VStack
-                Spacer()
-                NavigationLink {
-                    RegisterNicknameView(isFirstLaunching: $isFirstLaunching)
-                        .navigationTitle("")
-                } label: {
-                    ZStack {
-                        RoundedRectangle(cornerRadius: 12)
-                            .foregroundColor(.white)
-                        
-                        Text("다음")
-                            .font(.title2.bold())
-                            .foregroundColor(.main)
-                    }
-                    .frame(height: 60)
                 }
                 Spacer()
+                
             }// VStack
-            .padding(.top, 170)
-            .padding(.horizontal, 20)
-            .background(Color.main.ignoresSafeArea())
-            
-        }// ZStack
+            Spacer()
+            NavigationLink {
+                RegisterNicknameView(isFirstLaunching: $isFirstLaunching)
+                    .navigationTitle("")
+            } label: {
+                ZStack {
+                    RoundedRectangle(cornerRadius: 12)
+                        .foregroundColor(.white)
+                    
+                    Text("다음")
+                        .font(.title2.bold())
+                        .foregroundColor(.main)
+                }
+                .frame(height: 60)
+            }
+            Spacer()
+        }// VStack
+        .padding(.top, 60)
+        .padding(.horizontal, 20)
+        .background(Color.main.ignoresSafeArea())
+        
     }// body
 }// OnboardingLastView
 
 struct OnboardingLastView_Previews: PreviewProvider {
-
+    
     @State static var isFirstLaunching: Bool = true
     
     static var previews: some View {
